src/agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aws-lambda-powertools>=3.24.0",
    "aws-opentelemetry-distro>=0.15.0",
    "bedrock-agentcore>=1.2.1",
    "feedparser>=6.0.12",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "strands-agents>=1.25.0",
    "strands-agents-tools>=0.2.20",
    "tavily>=1.1.0",
]
